requirements.txt
torch>=1.10.0
transformers>=4.18.0
datasets>=2.0.0
evaluate>=0.3.0
hydra-core>=1.2.0
omegaconf>=2.2.0
numpy>=1.20.0
scikit-learn>=1.0.0
tensorboard>=2.8.0
tqdm>=4.64.0